+++ CHANGELOG.md
@@ -5,12 +5,17 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 
 ## [2.4.2] - 2022-07-18
+- Removed the -R option as this option is not available in Mongo 4.2.
+
+### Fixed
+- status: Moved json.dumps call to after Mongo database insert as json.dumps was converting "none" to "null".
+
 ### Changed
 - status: Removed checks on data type for Mongo insertion and refactored function.
 - process_request: Removed "use_arg" and "use_uri" from mongo_class.DB instance call.
 
-### Fixed
-- status: Moved json.dumps call to after Mongo database insert as json.dumps was converting "none" to "null".
+### Removed
+- run_repair, repair_db: The "repairDatabase" command is longer available in Mongo 4.2.
 
 
 ## [2.4.1] - 2022-06-28